refactor: MenuGroup 분리 및 의존성 제거
