+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -8,8 +8,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import javax.validation.constraints.NotNull;
 
 @Entity
@@ -28,30 +26,28 @@ public class Menu {
     @Embedded
     private Money price;
     @NotNull
-    @ManyToOne
-    @JoinColumn(name = "menu_group_id")
-    private MenuGroup menuGroup;
+    private Long menuGroupId;
     @Embedded
     private MenuProducts menuProducts;
 
     protected Menu() {
     }
 
-    private Menu(final Long id, final Name name, final Money price, final MenuGroup menuGroup, final MenuProducts menuProducts) {
+    private Menu(final Long id, final Name name, final Money price, final Long menuGroupId, final MenuProducts menuProducts) {
         this.id = id;
         this.name = name;
         this.price = price;
-        this.menuGroup = menuGroup;
+        this.menuGroupId = menuGroupId;
         validateMenuProductsSize(menuProducts);
         validateMenuPrice(price, menuProducts.calculateSum());
         this.menuProducts = menuProducts;
     }
 
-    public static Menu of(final Name name, final Money price, final MenuGroup menuGroup, final MenuProducts menuProducts) {
+    public static Menu of(final Name name, final Money price, final Long menuGroup, final MenuProducts menuProducts) {
         return new Menu(null, name, price, menuGroup, menuProducts);
     }
 
-    public static Menu of(final String name, final long price, final MenuGroup menuGroup, final MenuProducts menuProducts) {
+    public static Menu of(final String name, final long price, final Long menuGroup, final MenuProducts menuProducts) {
         return new Menu(null, Name.of(name), Money.valueOf(price), menuGroup, menuProducts);
     }
 
@@ -79,8 +75,8 @@ public class Menu {
         return price;
     }
 
-    public MenuGroup getMenuGroup() {
-        return menuGroup;
+    public Long getMenuGroupId() {
+        return menuGroupId;
     }
 
     public MenuProducts getMenuProducts() {
