+++ src/main/java/kitchenpos/domain/menugroup/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.menu;
+package kitchenpos.domain.menugroup;
 
 import org.springframework.data.jpa.repository.JpaRepository;
