+++ src/test/java/kitchenpos/application/integration/ApplicationIntegrationTest.java
@@ -2,16 +2,16 @@ package kitchenpos.application.integration;
 
 import kitchenpos.application.ProductService;
 import kitchenpos.application.TableGroupService;
-import kitchenpos.application.menu.MenuGroupService;
+import kitchenpos.application.mengroup.MenuGroupService;
 import kitchenpos.application.menu.MenuService;
 import kitchenpos.application.order.OrderService;
 import kitchenpos.application.table.TableService;
 import kitchenpos.domain.common.Money;
-import kitchenpos.dto.menu.CreateMenuGroupRequest;
 import kitchenpos.dto.menu.CreateMenuRequest;
-import kitchenpos.dto.menu.MenuGroupResponse;
 import kitchenpos.dto.menu.MenuProductRequest;
 import kitchenpos.dto.menu.MenuResponse;
+import kitchenpos.dto.menugroup.CreateMenuGroupRequest;
+import kitchenpos.dto.menugroup.MenuGroupResponse;
 import kitchenpos.dto.order.CreateOrderRequest;
 import kitchenpos.dto.order.OrderLineItemRequest;
 import kitchenpos.dto.order.OrderResponse;
